+++ cmd/ticket.go
@@ -69,7 +69,7 @@ var ticketCmd = &cobra.Command{
 				log.Fatalf("unexpected error '%v' reading url '%v'", err, url)
 			}
 			packageId := linkParts.PackageCode
-			err = sendsafely.DownloadFilesFromPackage(packageId, linkParts.KeyCode, C, filepath.Join("tickets", ticketId))
+			err = sendsafely.DownloadFilesFromPackage(packageId, linkParts.KeyCode, C, filepath.Join("tickets", ticketId, packageId))
 			if err != nil {
 				log.Fatal(err)
 			}