+++ changelog.md
@@ -1,7 +1,14 @@
+v0.2.3
+------
+* use package name for prefix
+* was unable to combine especially large files due to error in logic
+* more testing!!!
+
 v0.2.2
 ------
 * had init check for prompt backwords now works
 * updated docs and help to show subdmain
+
 v0.2.1
 ------
 * require init to have certain parameters